refactor: move bolt-action to 'core/elements' directory
