+++ packages/components/bolt-link/src/link.js
@@ -1,6 +1,6 @@
 import { props, define } from '@bolt/core/utils';
 import { html, render } from '@bolt/core/renderers/renderer-lit-html';
-import { BoltAction } from '@bolt/core/renderers/bolt-action';
+import { BoltAction } from '@bolt/core/elements/bolt-action';
 
 import classNames from 'classnames/bind';
